// src/app/pages/private/CreateNew/AddActivity/AddActivity.type.ts
export interface IValues {
  name: string;
  address: string;
  street: string;
  houseNumber: string;
  ward: string;
  district: string;
  city: string;
  country: string;
  price: number;
  introduce: string;
  startDate: string;
  fromHour: string;
  endDate: string;
  toHour: string;
  files: any;
  locationId: any;
  lat: number;
  lng: number;
}

export interface IProps {
  history: any;
}